requirements.txt
django
djangorestframework
celery
redis
celery
django_celery_beat
django_celery_results
python-decouple
djangorestframework_simplejwt
drf-yasg
django-filter==25.1